Single workspace configuration
