+++ .vscode/settings.json
@@ -1,5 +1,11 @@
 {
-    "files.watcherExclude": {
-        "**/target/**": true
-    }
+  "files.watcherExclude": {
+    "**/target/**": true
+  },
+  "[rust]": {
+    "editor.defaultFormatter": "rust-lang.rust-analyzer",
+    "editor.formatOnType": true,
+    "editor.formatOnPaste": true,
+    "editor.formatOnSave": true
+  }
 }
