+++ .vscode/tasks.json
@@ -1,31 +1,38 @@
 {
-    // See https://go.microsoft.com/fwlink/?LinkId=733558 
-    // for the documentation about the tasks.json format
-    "version": "2.0.0",
-    "tasks": [
-        {
-            /*
-             * This is the default cargo build task,
-             * but we need to provide a label for it,
-             * so we can invoke it from the debug launcher.
-             */
-            "label": "cargo_build",
-            "type": "process",
-            "command": "cargo",
-            "args": [
-                "build",
-                "--target",
-                "${config:rust-analyzer.cargo.target}",
-                "--features",
-                "${config:rust-analyzer.cargo.features}"
-            ],
-            "problemMatcher": [
-                "$rustc",
-            ],
-            "group": {
-                "kind": "build",
-                "isDefault": true
-            }
-        }
-    ]
+  // See https://go.microsoft.com/fwlink/?LinkId=733558
+  // for the documentation about the tasks.json format
+  "version": "2.0.0",
+  "tasks": [
+    {
+      /*
+       * This is the default cargo build task,
+       * but we need to provide a label for it,
+       * so we can invoke it from the debug launcher.
+       */
+      "label": "cargo_build",
+      "type": "shell",
+      "command": "cargo",
+      "args": [
+        "build",
+        "--bin",
+        "${fileBasenameNoExtension}",
+        "--features",
+        "${fileBasenameNoExtension}",
+        "--target",
+        "${fileDirnameBasename}"
+      ],
+      "problemMatcher": ["$rustc"],
+      "group": {
+        "kind": "build",
+        "isDefault": true
+      },
+      "presentation": {
+        "revealProblems": "onProblem",
+        "reveal": "silent",
+        "showReuseMessage": false,
+        "clear": true,
+        "close": false
+      }
+    }
+  ]
 }
