+++ README.md
@@ -6,6 +6,8 @@ The application can be run on mulitple architectures, and is controlled by condi
 
 - `STM32H745ZITx`
   - ARM Cortex-M7 core on a ST Nucleo H745ZI-Q board
+- `STM32C031C6Tx`
+  - ARM Cortex-M6 core on a STM32C0316-DK board
 - `nRF52833_xxAA`
   - ARM Cortex-M4 on a Micro:Bit v2 board
 - `esp32c3`
@@ -17,19 +19,8 @@ The application can be run on mulitple architectures, and is controlled by condi
 
 Use the **VSCode probe-rs-debug extension** for `probe-rs-debugger`
 
-- Edit the applicable `<chipname>.code-workspace` file, and adjust the locations marked with a **// CONFIGURE:** comment.
-- The `.vscode/launch.json` and `.vscode/tasks.json` are preconfigured, and will adjust to the values configured in the `<chipname>.code-workspace` file.
-- Open the applicable workspace file with 'Open Workspace from File ...', and you should be ready to debug.
+- The `.vscode/launch.json` and `.vscode/tasks.json` are preconfigured, and will adjust behaviour based on the active source file in the editor. The configuration uses VSCode variables referencing both the file name and the parent folder name, to determine the correct values in the configuration to use.
 
 ## Adding support for new chips
 
-Support for new chips, can be added by making the following modifications, using the appropriate chip name from `probe-rs-debugger list-chips`
-
-1. Copy one of the existing VSCode workspace files to an new file named: `<chipname>.code_workspace` - Update all locations marked with a **// CONFIGURE:** comment.
-   2.Edit the `Cargo.toml` file: - Add a [feature] for your chip. - If your feature uses a different [target.\<triple>] , then create a new dependency for it, using the existing target platforms as a template.
-2. Edit the `src/main.rs` file:
-   - Add your chip [feature] to the existing conditional compile structures, or add new ones if your chip requires custom crates.
-3. Add the appropriate linker file to the `linker_files` folder.
-   - Note, it must be named "<chip_name>\_memory.x"
-4. Add the appropriate CMSIS-SVD file to the `svd_files` folder.
-   - Note, it must be named "<chip_name>.svd"
+Support for new chips, can be added by making the copying one of the existing source files, and renaming it using the appropriate chip name from `probe-rs-debugger list-chips`. The new file should be placed in a folder named after the target-triple. Finally, `Cargo.toml` and `build.rs` should be updated to include the new chip name and features.
