--- RP2040.code-workspace
@@ -1,17 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "."
-        }
-    ],
-    "settings": {
-        "probe-rs-chipname": "RP2040", // CONFIGURE: Replace target feature
-        "rust-analyzer.cargo.features": [
-            "RP2040", // CONFIGURE: Replace target feature
-        ],
-        "probe-rs-vid-pid": "cafe:4005", // CONFIGURE: Replace with your probe's VID:PID
-        "rust-analyzer.checkOnSave.command": "clippy",
-        "rust-analyzer.cargo.target": "thumbv6m-none-eabi", // CONFIGURE: Replace target identity
-        "editor.formatOnSave": true
-    }
-}
