--- esp32c3.code-workspace
@@ -1,17 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "."
-        }
-    ],
-    "settings": {
-        "probe-rs-chipname": "esp32c3", // CONFIGURE: Replace target feature
-        "rust-analyzer.cargo.features": [
-            "esp32c3", // CONFIGURE: Replace target feature
-        ],
-        "probe-rs-vid-pid": "303a:1001", // CONFIGURE: Replace with your probe's VID:PID
-        "rust-analyzer.checkOnSave.command": "clippy",
-        "rust-analyzer.cargo.target": "riscv32imac-unknown-none-elf", // CONFIGURE: Replace target identity
-        "editor.formatOnSave": true,
-    }
-}
