--- nrf52833.code-workspace
@@ -1,17 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "."
-        }
-    ],
-    "settings": {
-        "probe-rs-chipname": "nRF52833_xxAA", // CONFIGURE: Replace target feature
-        "rust-analyzer.cargo.features": [
-            "nRF52833_xxAA", // CONFIGURE: Replace target feature
-        ],
-        "probe-rs-vid-pid": "0d28:0204", // CONFIGURE: Replace with your probe's VID:PID
-        "rust-analyzer.checkOnSave.command": "clippy",
-        "rust-analyzer.cargo.target": "thumbv7em-none-eabihf", // CONFIGURE: Replace target identity
-        "editor.formatOnSave": true
-    }
-}
