--- stm32h745.code-workspace
@@ -1,17 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {
-        "probe-rs-chipname": "STM32H745ZITx", // CONFIGURE: Replace target feature
-        "rust-analyzer.cargo.features": [
-            "STM32H745ZITx", // CONFIGURE: Replace target feature
-        ],
-        "probe-rs-vid-pid": "0483:374e", // CONFIGURE: Replace with your probe's VID:PID
-		"rust-analyzer.checkOnSave.command": "clippy",
-		"rust-analyzer.cargo.target": "thumbv7em-none-eabihf", // CONFIGURE: Replace target identity
-		"editor.formatOnSave": true
-	}
-}
